requirements.txt
selenium
python-dotenv
psycopg2
requests
